Fix: FollowRepository Following, Followed 수 조회하는 메서드 반환형 변경 Long -> Integer

+++ src/main/java/site/dittotrip/ditto_trip/follow/repository/FollowRepository.java
@@ -15,8 +15,8 @@ public interface FollowRepository extends JpaRepository<Follow, Long> {
 
     List<Follow> findByFollowedUser(User followedUser);
 
-    Long countByFollowingUser(User followingUser);
+    Integer countByFollowingUser(User followingUser);
 
-    Long countByFollowedUser(User followedUser);
+    Integer countByFollowedUser(User followedUser);
 
 }
